Rename shared package to common
